ListInfo.copy_to: RRR
First iteration of the duplicate-info missing API - we lived without
it for a while we will keep doing so - as implementing it goes through
lots of complexities like copying the dicts of the infos etc - see
installer copy. We will need to eradicate table first (absorb as info
attributes) then iterate mro pseudo-slots. Later (actually there was
such a high demand that table was eradicated but we still need some
work to have a proper copying API). Note I inlined the only use of
_mark_unchanged but that is # 241 territory (and ghosts brrr).

Absorb _do_copy

FFF copy_to