Remove remove_section:
The only use was in _games_lo.py in combination with saveSettings - the
latter can easily omit a section (and possible repetitions). Apart from
decluttering the API (new and sleek edition) drops the last use of
_parse_setting outside of _check_deleted - we don't support sections
for OBSEIniFile (note reSection is not in formatRes) and seeing the only
use has nothing to do with OBSE inis let's quickly drop it before it
becomes a thing. Iterating the ini and parsing the lines is almost
completely centralized now - see also next commit.