Inline _get_ci_settings: TTT
The key difference is the missing-section handling - what the
defaultSection is about. Since the inis we parse are supposed to be
read by the game (or xSE, but these have no sections) we should check
if the game reads settings outside of any section and if not drop
handling missing section entirely. We do treat missing-section settings
as comments - see next commit.

Note in OBSEIniFile._get_ci_settings we were iterating iniFile.readlines,
we now use read_ini_contents (so splitlines(keepends=False)) - should not
make a difference as we strip in _parse_setting.

Note also we had to keep the value for deleted settings also - it's not
needed but simplifies the code (and debugging potentially).