WIP fixups for deleted files handling and absorb applyTweakFile to base: EEE
EEE see 169d8347c1e4f3a3f6d696d4a660f89987f6bffc
EEE DefaultLowerDict(set)?
EEE use sect in try block only

The slight repetition in _fmt_setting could be absorbed in saveSettings
if we drop the OBSE override ;) The former could be added parameters to
format the section and deleted settings too but that would be baroque.
Note that if the wizard knew what kind of ini we are editing
generateTweakLines could go - that'd be a wonder. Note that
this is the last use of DefaultLowerDict and no setdefault remain.